Manual merge with refactor for code coverage
