+++ data-prepper-core/src/test/java/com/amazon/situp/parser/PipelineParserTests.java
@@ -13,10 +13,10 @@ import static com.amazon.situp.TestDataProvider.VALID_MULTIPLE_PIPELINE_CONFIG_F
 import static com.amazon.situp.TestDataProvider.VALID_MULTIPLE_PIPELINE_NAMES;
 import static com.amazon.situp.TestDataProvider.VALID_MULTIPLE_PROCESSORS_CONFIG_FILE;
 import static com.amazon.situp.TestDataProvider.VALID_MULTIPLE_SINKS_CONFIG_FILE;
+import static com.amazon.situp.TestDataProvider.VALID_SINGLE_PIPELINE_EMPTY_SOURCE_PLUGIN_FILE;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 public class PipelineParserTests {
@@ -28,10 +28,17 @@ public class PipelineParserTests {
         assertThat(actualPipelineMap.keySet(), is(VALID_MULTIPLE_PIPELINE_NAMES));
     }
 
+    @Test
+    public void testValidSinglePipelineEmptySourceSettings() {
+        final PipelineParser pipelineParser = new PipelineParser(VALID_SINGLE_PIPELINE_EMPTY_SOURCE_PLUGIN_FILE);
+        final Map<String, Pipeline> actualPipelineMap = pipelineParser.parseConfiguration();
+        assertThat(actualPipelineMap.keySet().size(), is(1));
+    }
+
     @Test //not preferring expected to validate exception message
     public void testCyclesInMultiplePipelines() {
         final PipelineParser pipelineParser = new PipelineParser(CYCLE_MULTIPLE_PIPELINE_CONFIG_FILE);
-        try{
+        try {
             pipelineParser.parseConfiguration();
         } catch (RuntimeException ex) {
             assertThat(ex.getMessage(), is("Provided configuration results in a loop, check pipeline: test-pipeline-1"));
@@ -41,7 +48,7 @@ public class PipelineParserTests {
     @Test //not preferring expected to validate exception message
     public void testInCorrectSourceMappingInMultiPipelines() {
         final PipelineParser pipelineParser = new PipelineParser(INCORRECT_SOURCE_MULTIPLE_PIPELINE_CONFIG_FILE);
-        try{
+        try {
             pipelineParser.parseConfiguration();
         } catch (RuntimeException ex) {
             assertThat(ex.getMessage(), is("Invalid configuration, expected source test-pipeline-1 for pipeline test-pipeline-2 is missing"));
@@ -51,7 +58,7 @@ public class PipelineParserTests {
     @Test
     public void testMissingNameInPipelineTypeInMultiPipelines() {
         final PipelineParser pipelineParser = new PipelineParser(MISSING_NAME_MULTIPLE_PIPELINE_CONFIG_FILE);
-        try{
+        try {
             pipelineParser.parseConfiguration();
         } catch (RuntimeException ex) {
             assertThat(ex.getMessage(), is("name is a required attribute for sink pipeline plugin, " +
@@ -62,7 +69,7 @@ public class PipelineParserTests {
     @Test
     public void testMissingPipelineTypeInMultiPipelines() {
         final PipelineParser pipelineParser = new PipelineParser(MISSING_PIPELINE_MULTIPLE_PIPELINE_CONFIG_FILE);
-        try{
+        try {
             pipelineParser.parseConfiguration();
         } catch (RuntimeException ex) {
             assertThat(ex.getMessage(), is("Invalid configuration, no pipeline is defined with name test-pipeline-4"));
@@ -83,7 +90,7 @@ public class PipelineParserTests {
 
     @Test
     public void testIncorrectConfigurationFilePath() {
-        try{
+        try {
             final PipelineParser pipelineParser = new PipelineParser("file_does_no_exist.yml");
             pipelineParser.parseConfiguration();
         } catch (ParseException ex) {
