+++ data-prepper-core/src/test/resources/single_pipeline_valid_empty_source_plugin_settings.yml
