+++ data-prepper-core/src/test/resources/valid_multiple_pipeline_configuration.yml
@@ -2,7 +2,7 @@
 test-pipeline-1:
   source:
     file:
-      someProperty: "someValue"
+      path: "/tmp/file-source.tmp"
   buffer:
     bounded_blocking: #to check non object nodes for plugins
   sink:
@@ -21,4 +21,4 @@ test-pipeline-3:
       name: "test-pipeline-2"
   sink:
     - file:
-       someProperty: "someValue"
+       path: "/tmp/todelete.txt"
