+++ data-prepper-plugins/common/src/test/java/com/amazon/situp/plugins/processor/ProcessorFactoryTests.java
@@ -9,23 +9,26 @@ import org.junit.Test;
 import java.util.HashMap;
 
 import static com.amazon.situp.plugins.PluginFactoryTests.NON_EXISTENT_EMPTY_CONFIGURATION;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertThrows;
 
 @SuppressWarnings("rawtypes")
 public class ProcessorFactoryTests {
-
+    private static String TEST_PIPELINE = "test-pipeline";
     /**
      * Tests if ProcessorFactory is able to retrieve default Source plugins by name
      */
     @Test
     public void testNewSinkClassByNameThatExists() {
         final PluginSetting noOpProcessorConfiguration = new PluginSetting("no-op", new HashMap<>());
+        noOpProcessorConfiguration.setPipelineName(TEST_PIPELINE);
         final Processor actualProcessor = ProcessorFactory.newProcessor(noOpProcessorConfiguration);
         final Processor expectedProcessor = new NoOpProcessor();
-        assertNotNull(actualProcessor);
-        assertEquals(expectedProcessor.getClass().getSimpleName(), actualProcessor.getClass().getSimpleName());
+        assertThat(actualProcessor, notNullValue());
+        assertThat(actualProcessor.getClass().getSimpleName(), is(equalTo(expectedProcessor.getClass().getSimpleName())));
     }
 
     /**
