+++ data-prepper-plugins/common/src/test/java/com/amazon/situp/plugins/source/StdInSourceTests.java
