+++ data-prepper-plugins/common/src/test/resources/test-file-source.tst
